+++ CHANGES.md
@@ -5,6 +5,7 @@ List changes at major/minor level. Patches should be straightforward.
 # 2.0.0
 - Stop adding breaking changes in minor versions.
 - Drop RXJS, use ol/native mechanisms instead.
+- Remove the UnByKeyAll method, you can use the native ol/Observable.unByKeyAll method instead.
 
 # 1.4.0
 - Remove lodash.